VZ-5631 End to end tests for kube-state-metrics, consolidate all e2e tests for Prometheus stack into one suite
