--- tests/e2e/verify-install/prometheusoper/prometheusoper_suite_test.go
@@ -1,16 +1,0 @@
-// Copyright (c) 2022, Oracle and/or its affiliates.
-// Licensed under the Universal Permissive License v 1.0 as shown at https://oss.oracle.com/licenses/upl.
-
-package prometheusoper
-
-import (
-	"testing"
-
-	"github.com/onsi/ginkgo/v2"
-	"github.com/onsi/gomega"
-)
-
-func TestPrometheusOperator(t *testing.T) {
-	gomega.RegisterFailHandler(ginkgo.Fail)
-	ginkgo.RunSpecs(t, "Prometheus Operator Suite")
-}
