+++ tests/e2e/verify-install/promstack/promstack_suite_test.go
@@ -1,7 +1,7 @@
 // Copyright (c) 2022, Oracle and/or its affiliates.
 // Licensed under the Universal Permissive License v 1.0 as shown at https://oss.oracle.com/licenses/upl.
 
-package promadapter
+package promstack
 
 import (
 	"testing"
@@ -10,7 +10,7 @@ import (
 	"github.com/onsi/gomega"
 )
 
-func TestPrometheusAdapter(t *testing.T) {
+func TestPrometheusStack(t *testing.T) {
 	gomega.RegisterFailHandler(ginkgo.Fail)
-	ginkgo.RunSpecs(t, "Prometheus Adapter Suite")
+	ginkgo.RunSpecs(t, "Prometheus Stack Suite")
 }
